Edited update and delete controller methods and views

+++ ContosoUniversity/Controllers/StudentsController.cs
@@ -73,7 +73,7 @@ namespace ContosoUniversity.Controllers
             {
                 //Log the error (uncomment ex variable name and write log)
                 ModelState.AddModelError("", "Unable to save changes" +
-                    "Try again, and it if the problem persists " +
+                    "Try again and it if the problem persists " +
                     "see your system administrator");
             }
 
@@ -108,31 +108,33 @@ namespace ContosoUniversity.Controllers
                 return NotFound();
             }
 
-            if (ModelState.IsValid)
+            var studentToUpdate = await _context.Student.FirstOrDefaultAsync(s => s.Id == id);
+
+            if (await TryUpdateModelAsync<Student>(
+                    studentToUpdate,
+                    "",
+                    s => s.FirstName, s => s.LastName, s => s.EnrollmentDate))
             {
                 try
                 {
-                    _context.Update(student);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!StudentExists(student.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error (uncomment ex variable name and write log)
+                    ModelState.AddModelError("", "Unable to save changes" +
+                        "Try again and it if the problem persists " +
+                        "see your system administrator");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(student);
+
+            return View(studentToUpdate);
+
         }
 
         // GET: Students/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
@@ -140,10 +142,19 @@ namespace ContosoUniversity.Controllers
             }
 
             var student = await _context.Student
+                .AsNoTracking()
                 .FirstOrDefaultAsync(m => m.Id == id);
+            
             if (student == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                    "Delete failed. Try again and if the problems persists " +
+                    "see your system administrator.";
             }
 
             return View(student);
@@ -155,9 +166,22 @@ namespace ContosoUniversity.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var student = await _context.Student.FindAsync(id);
-            _context.Student.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            
+            if(student == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Student.Remove(student);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /**/)
+            {
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool StudentExists(int id)
